feat: split server/tools_ci.py and workspace/clone.py into submodules
Splits tools_ci.py into tools_ci_watch.py (wait_for_ci, get_ci_status,
_auto_trigger_ci) and tools_ci_merge_queue.py (toggle_auto_merge, enqueue_pr,
wait_for_merge_queue). Splits clone.py into _clone_detect.py (detection
helpers + URL classification) and _clone_remote.py (remote probing and
isolation). All readOnlyHint annotations preserved. Public surfaces stable.
