+++ src/autoskillit/workspace/clone.py
@@ -18,318 +18,25 @@ cross-user hardlink risk (CVE-2024-32020).
 import os
 import shutil
 import subprocess
-from dataclasses import dataclass
 from datetime import datetime
 from pathlib import Path
 from typing import Literal
 
 from autoskillit.core import GENERATED_FILES, CloneResult, get_logger, is_protected_branch
+from autoskillit.workspace._clone_detect import (
+    RUNS_DIR,
+    classify_remote_url,
+    detect_branch,
+    detect_source_dir,
+    detect_uncommitted_changes,
+    detect_unpublished_branch,
+)
+from autoskillit.workspace._clone_remote import (
+    _ensure_origin_isolated,
+    _probe_clone_source_url,
+)
 
 logger = get_logger(__name__)
-
-RUNS_DIR = "autoskillit-runs"
-
-# URL prefixes that unambiguously identify a network remote
-_NETWORK_URL_PREFIXES = ("https://", "http://", "git@", "git://", "ssh://", "file://")
-
-
-def _is_not_file_url(url: str) -> bool:
-    """Return True if url is a usable clone source (not a self-referential file:// URL).
-
-    file:// URLs are written by _ensure_origin_isolated to point at the clone directory
-    itself. Using them as a clone source would clone from the stale local working tree
-    instead of fetching fresh state from the real remote (the #817 bug).
-
-    Local bare paths (no scheme) and real network URLs (https://, git@, etc.) are both
-    valid clone sources and are accepted.
-    """
-    return bool(url) and not url.startswith("file://")
-
-
-def classify_remote_url(url: str) -> str:
-    """Classify a git remote URL as 'network', 'bare_local', 'nonbare_local', 'none', or 'unknown'.
-
-    Network URLs: https://, http://, git@, git://, ssh://, file:// scheme.
-    Bare local: a local filesystem path where git rev-parse --is-bare-repository returns true.
-    Nonbare local: a local filesystem path that is a non-bare git repo.
-    None: empty string (no remote configured).
-    Unknown: local path that does not exist or is not a git repo.
-    """
-    if not url:
-        return "none"
-    if any(url.startswith(prefix) for prefix in _NETWORK_URL_PREFIXES):
-        return "network"
-    # Treat as a local filesystem path
-    path = Path(url).expanduser().resolve()
-    if not path.exists():
-        return "unknown"
-    result = subprocess.run(
-        ["git", "-C", str(path), "rev-parse", "--is-bare-repository"],
-        capture_output=True,
-        text=True,
-    )
-    if result.returncode != 0:
-        return "unknown"
-    return "bare_local" if result.stdout.strip() == "true" else "nonbare_local"
-
-
-def detect_source_dir(cwd: str) -> str:
-    """Detect the git repository root for cwd, falling back to cwd.
-
-    Shells 'git rev-parse --show-toplevel' from cwd. Returns cwd unchanged
-    if not in a git repository or if git is not available.
-    """
-    _cwd = cwd if cwd else str(Path.cwd())
-    result = subprocess.run(
-        ["git", "rev-parse", "--show-toplevel"],
-        cwd=_cwd,
-        capture_output=True,
-        text=True,
-    )
-    if result.returncode == 0:
-        return result.stdout.strip()
-    return _cwd
-
-
-def detect_branch(source_dir: str) -> str:
-    """Detect the current HEAD branch in source_dir.
-
-    Returns the branch name on success. Returns "" on git failure.
-    Returns the literal "HEAD" when the repo is in detached HEAD state;
-    callers must treat "HEAD" as no usable branch name.
-    """
-    result = subprocess.run(
-        ["git", "rev-parse", "--abbrev-ref", "HEAD"],
-        cwd=source_dir,
-        capture_output=True,
-        text=True,
-    )
-    if result.returncode == 0:
-        return result.stdout.strip()
-    return ""
-
-
-def detect_uncommitted_changes(source_dir: str) -> list[str]:
-    """Return porcelain status lines when uncommitted changes exist.
-
-    Returns empty list when the working tree is clean or when git is
-    unavailable. Never raises.
-    """
-    result = subprocess.run(
-        ["git", "status", "--porcelain"],
-        cwd=source_dir,
-        capture_output=True,
-        text=True,
-    )
-    if result.returncode == 0:
-        return [line for line in result.stdout.splitlines() if line.strip()]
-    return []
-
-
-def detect_unpublished_branch(source_dir: str, branch: str) -> bool:
-    """Return True if `branch` has no ref on `origin` in `source_dir`.
-
-    Fail-open: returns False (do not block) when:
-    - No 'origin' remote is configured
-    - Any git command errors (network issue, non-git dir, etc.)
-    - Network probe times out (firewalled remote, unreachable SSH host, etc.)
-    Returns True only when origin is reachable and explicitly has no matching ref.
-    """
-    remote_check = subprocess.run(
-        ["git", "remote", "get-url", "origin"],
-        cwd=source_dir,
-        capture_output=True,
-        text=True,
-    )
-    if remote_check.returncode != 0:
-        return False  # No remote configured — can't confirm, don't block
-
-    try:
-        ls_remote = subprocess.run(
-            ["git", "ls-remote", "--exit-code", "origin", f"refs/heads/{branch}"],
-            cwd=source_dir,
-            capture_output=True,
-            text=True,
-            timeout=15,
-        )
-    except subprocess.TimeoutExpired:
-        return False  # Network hung — fail-open, don't block the pipeline
-    # exit code 2 from --exit-code means "no matching refs found"
-    return ls_remote.returncode == 2
-
-
-def _add_or_set_upstream(clone_path: Path, url: str) -> None:
-    """Add or update the upstream remote in the clone.
-
-    Handles the case where upstream already exists (clone_local copies .git as-is
-    and the source may already have an upstream remote).
-    """
-    result = subprocess.run(
-        ["git", "remote", "add", "upstream", url],
-        cwd=str(clone_path),
-        capture_output=True,
-    )
-    if result.returncode != 0:
-        stderr = (
-            result.stderr.decode(errors="replace")
-            if isinstance(result.stderr, bytes)
-            else result.stderr
-        )
-        if "already exists" not in stderr:
-            raise RuntimeError(
-                f"git remote add upstream failed: {stderr.strip()}"
-                f"\nclone_path={clone_path}, url={url}"
-            )
-        # upstream already exists (e.g. clone_local copied it from source); update the URL
-        set_url = subprocess.run(
-            ["git", "remote", "set-url", "upstream", url],
-            cwd=str(clone_path),
-            capture_output=True,
-        )
-        if set_url.returncode != 0:
-            set_stderr = (
-                set_url.stderr.decode(errors="replace")
-                if isinstance(set_url.stderr, bytes)
-                else set_url.stderr
-            )
-            raise RuntimeError(
-                f"git remote set-url upstream failed: {set_stderr.strip()}"
-                f"\nclone_path={clone_path}, url={url}"
-            )
-
-
-def _ensure_origin_isolated(clone_path: Path, known_url: str) -> None:
-    """Rewrite origin to a self-referential file:// URL unconditionally.
-
-    Claude Code reads 'origin' to resolve the project root; a file:// URL
-    cannot match any registered GitHub project, so the clone is treated as
-    a fresh project rooted at clone_path — not aliased to the source repo.
-
-    This must fire for every successful clone regardless of URL availability.
-    The conditional guard that previously skipped this for clone_local and
-    empty-URL cases was the #377 compounding regression.
-
-    After setting origin, calls _add_or_set_upstream with known_url only
-    if known_url is truthy — the real remote URL is stored in 'upstream'
-    when available.
-    """
-    file_url = f"file://{clone_path}"
-    add_origin = subprocess.run(
-        ["git", "remote", "add", "origin", file_url],
-        cwd=str(clone_path),
-        capture_output=True,
-    )
-    if add_origin.returncode != 0:
-        add_stderr = (
-            add_origin.stderr.decode(errors="replace")
-            if isinstance(add_origin.stderr, bytes)
-            else add_origin.stderr
-        )
-        if "already exists" not in add_stderr:
-            raise RuntimeError(
-                f"git remote add origin failed: {add_stderr.strip()}\nclone_path={clone_path}"
-            )
-        # origin already exists (cloned from remote); rewrite it
-        set_origin = subprocess.run(
-            ["git", "remote", "set-url", "origin", file_url],
-            cwd=str(clone_path),
-            capture_output=True,
-        )
-        if set_origin.returncode != 0:
-            set_origin_stderr = (
-                set_origin.stderr.decode(errors="replace")
-                if isinstance(set_origin.stderr, bytes)
-                else set_origin.stderr
-            )
-            raise RuntimeError(
-                f"git remote set-url origin failed: {set_origin_stderr.strip()}"
-                f"\nclone_path={clone_path}"
-            )
-    if known_url:
-        _add_or_set_upstream(clone_path, known_url)
-
-
-@dataclass(frozen=True)
-class CloneSourceResolution:
-    """Result of probing a source directory for its clone-source remote URL.
-
-    Tries 'upstream' first (real remote when source_dir is a previous autoskillit
-    clone), then 'origin'. Returns the first non-file:// URL found.
-
-    reason:
-        "ok"        — a usable remote URL was found and returned
-        "no_origin" — no origin remote is configured in this repo
-        "timeout"   — subprocess.run timed out (30 s)
-        "error"     — subprocess returned non-zero exit code
-    url: the remote URL (empty string unless reason == "ok")
-    stderr: subprocess stderr (empty unless reason == "error")
-    """
-
-    reason: Literal["ok", "no_origin", "timeout", "error"]
-    url: str
-    stderr: str
-
-
-def _probe_single_remote(source: Path, remote_name: str) -> CloneSourceResolution:
-    """Run ``git remote get-url <remote_name>`` in source and return a typed result.
-
-    Shared by _probe_clone_source_url for each candidate remote name.
-    """
-    _no_remote_markers = ("no such remote",)
-    try:
-        result = subprocess.run(
-            ["git", "remote", "get-url", remote_name],
-            cwd=str(source),
-            capture_output=True,
-            text=True,
-            timeout=30,
-        )
-    except subprocess.TimeoutExpired:
-        return CloneSourceResolution(reason="timeout", url="", stderr="")
-    except OSError as exc:
-        return CloneSourceResolution(reason="error", url="", stderr=str(exc))
-
-    if result.returncode != 0:
-        stderr = result.stderr.strip()
-        if any(m in stderr.lower() for m in _no_remote_markers):
-            return CloneSourceResolution(reason="no_origin", url="", stderr=stderr)
-        return CloneSourceResolution(reason="error", url="", stderr=stderr)
-
-    url = result.stdout.strip()
-    if not url:
-        return CloneSourceResolution(reason="no_origin", url="", stderr="")
-    return CloneSourceResolution(reason="ok", url=url, stderr="")
-
-
-def _probe_clone_source_url(source: Path) -> CloneSourceResolution:
-    """Probe source for the best remote URL to use as the git clone source.
-
-    Avoids cloning from stale file:// local state:
-    1. Try 'upstream' — when source_dir is a previous autoskillit clone,
-       _ensure_origin_isolated sets upstream to the real remote and
-       origin to a self-referential file:// URL. Using upstream avoids cloning
-       from the stale local filesystem (fix for #817).
-    2. Try 'origin' — if upstream is absent or is itself a file:// URL, falls
-       back to origin, which is the real remote in a fresh dev checkout.
-    3. If neither yields a non-file:// URL, returns origin's result so callers
-       get the appropriate error (no_origin / error / local path as-is).
-
-    file:// URLs are explicitly excluded because _ensure_origin_isolated writes
-    them to point at the clone directory itself — a stale local reference.
-    Local bare paths (no scheme) and real network URLs are both accepted.
-    """
-    last_result: CloneSourceResolution | None = None
-    for remote_name in ("upstream", "origin"):
-        result = _probe_single_remote(source, remote_name)
-        last_result = result
-        if result.reason == "ok" and _is_not_file_url(result.url):
-            return result
-
-    # No non-file:// URL found — return the last probed result (origin) rather
-    # than re-invoking _probe_single_remote, which would be a redundant subprocess.
-    if last_result is not None:
-        return last_result
-    return _probe_single_remote(source, "origin")
 
 
 def clone_repo(
